+++ public/pages/Decoders/containers/DecoderFormPage.tsx
@@ -1,12 +1,7 @@
 import React, { useState, useEffect, useCallback } from 'react';
 import { NotificationsStart } from 'opensearch-dashboards/public';
 import { Form, Formik, FormikErrors } from 'formik';
-import {
-  decoderFormDefaultValue,
-  DecoderFormModel,
-  mapDecoderToForm,
-  mapYamlObjectToDecoder,
-} from '../components/mappers';
+import { decoderFormDefaultValue, mapYamlToLosslessDecoder } from '../components/mappers';
 import { YamlForm } from '../components/YamlForm';
 import {
   errorNotificationToast,
@@ -33,6 +28,7 @@ import {
 import { DecoderDocument } from '../../../../types/Decoders';
 import { DataStore } from '../../../store/DataStore';
 import { RouteComponentProps } from 'react-router-dom';
+import { validate } from 'joi';
 
 const editorTypes = [
   {
@@ -62,8 +58,8 @@ export const DecoderFormPage: React.FC<DecoderFormPageProps> = (props) => {
   const [isLoading, setIsLoading] = useState(false);
   const [selectedEditorType, setSelectedEditorType] = useState('yaml');
   const [integrationType, setIntegrationType] = useState<string>('');
-  const [decoder, setDecoder] = useState<DecoderDocument | undefined>(undefined);
-  const [initialValue, setInitialValue] = useState<DecoderFormModel>(decoderFormDefaultValue);
+  const [rawDecoder, setRawDecoder] = useState<string>(decoderFormDefaultValue);
+  const [decoder, setDecoder] = useState<DecoderDocument>();
 
   const { loading: loadingIntegrations, options: integrationTypeOptions } = useIntegrationSelector({
     notifications,
@@ -75,11 +71,9 @@ export const DecoderFormPage: React.FC<DecoderFormPageProps> = (props) => {
         setIsLoading(true);
         try {
           const response = await DataStore.decoders.getDecoder(idDecoder, spaceDecoder);
-          setDecoder(response?.document);
+          setRawDecoder(response?.decoder ?? decoderFormDefaultValue);
+          setDecoder(mapYamlToLosslessDecoder(response?.decoder ?? ''));
           setIntegrationType(response?.integrations?.[0] || '');
-          if (response?.document) {
-            setInitialValue(mapDecoderToForm(response.document));
-          }
           setBreadcrumbs([
             BREADCRUMBS.NORMALIZATION,
             BREADCRUMBS.DECODERS,
@@ -118,7 +112,7 @@ export const DecoderFormPage: React.FC<DecoderFormPageProps> = (props) => {
   }, []);
 
   const createDecoder = useCallback(
-    async (values: DecoderFormModel) => {
+    async (values: DecoderDocument) => {
       if (!values || !integrationType) {
         errorNotificationToast(
           notifications,
@@ -158,7 +152,7 @@ export const DecoderFormPage: React.FC<DecoderFormPageProps> = (props) => {
   );
 
   const updateDecoder = useCallback(
-    async (values: DecoderFormModel) => {
+    async (values: DecoderDocument) => {
       if (!values) {
         errorNotificationToast(notifications, 'retrieve', 'decoder', 'No decoder to update');
         return;
@@ -192,7 +186,7 @@ export const DecoderFormPage: React.FC<DecoderFormPageProps> = (props) => {
   );
 
   const handleOnClick = useCallback(
-    async (values: DecoderFormModel) => {
+    async (values: DecoderDocument) => {
       if (action === 'create') {
         await createDecoder(values);
       } else if (action === 'edit') {
@@ -202,14 +196,17 @@ export const DecoderFormPage: React.FC<DecoderFormPageProps> = (props) => {
     [action, createDecoder, updateDecoder]
   );
 
-  const validateForm = useCallback((values: DecoderFormModel) => {
-    const errors: FormikErrors<DecoderFormModel> = {};
+  const validateForm = useCallback((values: { rawDecoder: string }) => {
+    const errors: FormikErrors<DecoderDocument> = {};
 
-    if (!values.name) {
+    // FIXME: This is making a transformation on each detected change in the yaml form, this could create a lot of overhead
+    const decoder = mapYamlToLosslessDecoder(values.rawDecoder);
+
+    if (!decoder.name) {
       errors.name = 'Decoder name is required';
     }
 
-    const parts = values.name.split('/');
+    const parts = decoder.name.split('/');
 
     if (parts.length !== 3) {
       errors.name = "Decoder name must have exactly 3 parts 'decoder/<name>/<version>'";
@@ -246,13 +243,13 @@ export const DecoderFormPage: React.FC<DecoderFormPageProps> = (props) => {
       ) : (
         <Formik
           key={decoder?.id || 'new-decoder'}
-          initialValues={initialValue}
+          initialValues={{ rawDecoder: rawDecoder }}
           validateOnMount={true}
           enableReinitialize={true}
           validate={validateForm}
           onSubmit={(values, { setSubmitting }) => {
             setSubmitting(false);
-            handleOnClick(values);
+            handleOnClick(mapYamlToLosslessDecoder(values.rawDecoder));
           }}
         >
           {(props) => (
@@ -298,14 +295,13 @@ export const DecoderFormPage: React.FC<DecoderFormPageProps> = (props) => {
 
                 {selectedEditorType === 'yaml' && (
                   <YamlForm
-                    decoder={decoder ? decoder : props.values}
+                    rawDecoder={props.values.rawDecoder}
                     isInvalid={Object.keys(props.errors).length > 0}
                     errors={Object.keys(props.errors).map(
-                      (key) => props.errors[key as keyof DecoderFormModel] as string
+                      (key) => (props.errors as Record<string, string>)[key]
                     )}
                     change={(e) => {
-                      const formState = mapYamlObjectToDecoder(e);
-                      props.setValues(formState);
+                      props.setValues({ rawDecoder: e });
                     }}
                   />
                 )}