+++ types/Decoders.ts
@@ -8,7 +8,7 @@ import { CatalogResourceMetadata } from './ResourceMetadata';
 export type DecoderMetadata = CatalogResourceMetadata;
 
 export interface DecoderDocument {
-  id: string;
+  id?: string;
   name: string;
   enabled?: boolean;
   metadata: DecoderMetadata;